Add files via upload
Files to peform hat operations in angular velocity vectors, and conversion between the lie Algebra and the lie group of SE(3). Through the conventional use of the exponential and log map. A sample file with an example of how to generate and plot 3d objects is provided. In the future more work will be done to reduce the amount of lines necessary before a full working model is created